Ran and reviewed code.
Overall, such good work!!! It also felt amazing to use because of the tty-prompt gem, and I'm so glad to know it exists so I can use it later on. I would definitely say that there aren't enough commits to represent the timeline of how this code came to be, so setting yourself up to make regular commits would be a major plus for me.